demo/requirements.txt
flask_script
flask_fragment
flask_sqlalchemy
flask_login
flask_wtf
flask_debugtoolbar
